Update jboss-mode.el
Changed goto-next-* methods to make more sense: For messages of level DEBUG and TRACE i want to be able to jump from one message to the next with the same level. But for messages of level ERROR, INFO and WARN i want to stop at the the next message that has level lower than or equal to the current one.